2022.03.04: * fixed: src/utility/utility.cpp: UTF-8 support in the file system functions for `std::string` based paths * fixed: CMakeLists.txt: x64 build fix * changed: src: removed `std::move` calls on a return value to avoid RVO elision
git-svn-id: svn+ssh://svn.code.sf.net/p/tacklelib/tacklelib/trunk@546 ffccfebd-399a-4e7c-9cd2-5fb927d1fecb

+++ src/utility/memory.cpp
@@ -42,7 +42,7 @@ namespace utility {
             } BOOST_SCOPE_EXIT_END
 
             if (proc_id) {
-                proc_handle = OpenProcess(PROCESS_QUERY_INFORMATION, FALSE, proc_id);
+                proc_handle = OpenProcess(PROCESS_QUERY_INFORMATION, FALSE, DWORD(proc_id));
             }
             else {
                 proc_handle = GetCurrentProcess();
